+++ aws/xfce-xrdp-k8s-ec2-setup.sh
@@ -17,4 +17,5 @@ sudo reboot
 # Install Greencap K8s.
 git clone https://github.com/greencapk8s/greencap-k8s.git
 cd greencap-k8s
-USER_NAME_INSTALL="ubuntu" ./installers/run-installers.sh
+SETUP_TYPE="minimal" # minimal, full, custom
+sudo ./greencap.sh --local-debug --setup-type $SETUP_TYPE --user-name "ubuntu"