+++ installers/ingress-controller-install.sh
@@ -1,12 +1,11 @@
 #!/bin/bash
-# Script to install ingress controller.
 
 set -e
 
 INGRESS_PATH="./infra-code-manifests/ingress-nginx/ingress.yaml"
 
 echo "=========================================="
-echo "🔧 Installing ingress controller"
+echo "Installing ingress controller"
 echo "=========================================="
 
 kubectl apply -f "$INGRESS_PATH"
@@ -15,7 +14,4 @@ kubectl wait --namespace ingress-nginx \
   --selector=app.kubernetes.io/component=controller \
   --timeout=180s
 
-echo ""
-echo "=========================================="
-echo "✅ Ingress controller installed successfully!"
-echo "=========================================="
+echo "==> Ingress controller installed successfully!"