+++ installers/kube-dash-install.sh
@@ -10,7 +10,7 @@ echo "=========================================="
 echo "📝 Adding entries to /etc/hosts..."
 sudo bash -c 'echo "127.0.0.1 kubernetes-dashboard.greencap" >> /etc/hosts'
 
-kubectl apply -f https://raw.githubusercontent.com/kubernetes/dashboard/v2.7.0/aio/deploy/recommended.yaml
+kubectl apply -f ./infra-code-manifests/kubernetes-dashboard/kube-dash.yaml
 kubectl apply -f ./infra-code-manifests/kubernetes-dashboard/dash-admin.yaml
 kubectl -n kubernetes-dashboard create token admin-user; echo
 kubectl apply -f ./infra-code-manifests/kubernetes-dashboard/dash-ing.yaml