+++ installers/kubectl-top-install.sh
@@ -3,10 +3,8 @@
 set -e
 
 echo "=========================================="
-echo "Installing Kind and Creating Cluster"
+echo "Installing kubectl top"
 echo "=========================================="
-
-echo "📦 Installing kubectl top..."
 
 # Install metrics-server
 kubectl apply -f https://github.com/kubernetes-sigs/metrics-server/releases/latest/download/components.yaml
@@ -14,4 +12,4 @@ kubectl apply -f https://github.com/kubernetes-sigs/metrics-server/releases/late
 # Patch metrics-server to allow insecure TLS to kind.
 kubectl patch deployment metrics-server -n kube-system --type='json' -p='[{"op": "add", "path": "/spec/template/spec/containers/0/args/-", "value": "--kubelet-insecure-tls"}]'
 
-echo "✅ kubectl top installed successfully!"
+echo "==> kubectl top installed successfully!"